refactor: auth handler hook 통합, 로직 정리 (FE-95)
